requirements.txt
mcp>=1.1.0
upstash-vector>=0.5.0
groq>=0.33.0
python-dotenv>=1.0.0
